+++ Localization/deDE.PluginStrings.lua
@@ -1,4 +1,4 @@
--- [[ Exported at 2023-03-23 19-45-49 ]] --
+-- [[ Exported at 2023-03-24 00-17-00 ]] --
 -- [[ This code is automatically generated as an export from ]] --
 -- [[ an SQLite database and is not meant for manual edit. ]] --